# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorBody'
      success:
        type: boolean
    type: object
...
  request.CreateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        type: integer
      slug:
        description: 可选，为空则自动生成
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        minimum: 0
        type: integer
    required:
    - name
    type: object
  request.CreatePageRequest:
    properties:
      content:
        type: string
      cover_image:
        maxLength: 500
        type: string
      order:
        minimum: 0
        type: integer
      show_in_nav:
        type: boolean
      slug:
        maxLength: 200
        minLength: 1
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      summary:
        maxLength: 500
        type: string
      template:
        enum:
        - default
        - about
        - projects
        - footprints
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  request.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        type: integer
      slug:
        description: 可选，为空则保持不变
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        minimum: 0
        type: integer
    type: object
  request.UpdatePageRequest:
    properties:
      content:
        type: string
      cover_image:
        maxLength: 500
        type: string
      order:
        minimum: 0
        type: integer
      show_in_nav:
        type: boolean
      slug:
        maxLength: 200
        minLength: 1
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      summary:
        maxLength: 500
        type: string
      template:
        enum:
        - default
        - about
        - projects
        - footprints
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
      version:
        minimum: 1
        type: integer
    required:
    - version
    type: object
...
  service.ContributionsGroup:
    properties:
      contributions:
        additionalProperties:
          type: integer
        description: 'key: MM-DD, value: count'
        type: object
      year:
        type: integer
    type: object
  service.ContributionsResponse:
    properties:
      groups:
        description: 按年份分组
        items:
          $ref: '#/definitions/service.ContributionsGroup'
        type: array
      total:
        description: 过去一年总贡献数
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API 支持
    url: http://www.swagger.io/support
  description: 博客后端 API 服务，支持文章管理、配置文件认证等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 博客系统 API
  version: "1.0"
paths:
...
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取分类列表
      tags:
      - category
    post:
      consumes:
      - application/json
      parameters:
      - description: 分类数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建分类
      tags:
      - category
  /categories/{id}:
    delete:
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: 无内容
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除分类
      tags:
      - category
    get:
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据ID获取分类
      tags:
      - category
    put:
      consumes:
      - application/json
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新分类
      tags:
      - category
  /categories/slug/{slug}:
    get:
      parameters:
      - description: 分类Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据Slug获取分类
      tags:
      - category
  /github/contributions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ContributionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取 GitHub 贡献日历
      tags:
      - github
...
  /pages:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 按状态筛选
        in: query
        name: status
        type: string
      - description: 按模板筛选
        in: query
        name: template
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取页面列表
      tags:
      - page
    post:
      consumes:
      - application/json
      parameters:
      - description: 页面数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreatePageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建页面
      tags:
      - page
  /pages/{id}:
    delete:
      parameters:
      - description: 页面ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: 无内容
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除页面
      tags:
      - page
    get:
      parameters:
      - description: 页面ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据 ID 获取页面
      tags:
      - page
    put:
      consumes:
      - application/json
      parameters:
      - description: 页面ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页面数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新页面
      tags:
      - page
...
  /pages/slug/{slug}:
    get:
      parameters:
      - description: 页面 Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据 Slug 获取页面
      tags:
      - page
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证令牌，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"